update st venant-kirchhoff #2
should be correct now. also corrected bug in compressible neohookean. i was computing right cauchy greeen but it should have been left cauchy green. fixed now. Could be moved to blas, but was not sure how to do that. 3x3 matrices need to be convered to column vectors I think.